requirements.txt
mem0ai>=0.1.60
langchain>=0.1.0
langgraph>=0.0.10
pydantic>=2.5.0
typer>=0.9.0
python-dotenv>=1.0.0
pytest>=7.4.0
openai>=1.0.0
aiohttp>=3.9.0
black>=23.0.0
flake8>=6.1.0
tenacity>=8.2.0
rich>=13.0.0
pydantic-settings>=2.1.0
mcp[cli]>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
dagger-io>=0.9.0
dspy-ai>=2.0.0
